+++ src/utils/general/wishlist.ts
@@ -34,6 +34,11 @@ export async function addToWishlist(user: any, product: any) {
   const supabase = await createClient();
 
   try {
+    if (!user) {
+      console.error("User is not logged in.");
+      throw new Error("Please sign in to add fragrances to your wishlist.");
+    }
+
     const { data, error } = await supabase
       .from("wishlist")
       .insert({
@@ -42,7 +47,6 @@ export async function addToWishlist(user: any, product: any) {
       })
       .select(`*, product:product_id(*)`)
       .single();
-
     if (error) {
       console.error(error.message);
       throw new Error(
@@ -68,12 +72,18 @@ export async function removeFromWishlist(user: any, product: any) {
   const supabase = await createClient();
 
   try {
+    if (!user) {
+      console.error("User is not logged in.");
+      throw new Error(
+        "Please sign in to remove fragrances from your wishlist."
+      );
+    }
+
     const { data, error } = await supabase
       .from("wishlist")
       .delete()
       .eq("user_id", user.id)
       .eq("product_id", product.id);
-
     if (error) {
       console.error(error.message);
       throw new Error(
